service/Services/entities/src/main/java/org/infy/idp/entities/jobs/UserRolesPermissions.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.idp.entities.jobs;

import java.util.List;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store user role and permission
 * 
 * @author Infosys
 *
 */
public class UserRolesPermissions {

	@SerializedName("user_id")
	@Expose
	private String userId;
	@SerializedName("roles")
	@Expose
	private List<String> roles;
	@SerializedName("permissions")
	@Expose
	private List<String> permissions;
	@SerializedName("orgName")
	@Expose
	private String orgName;

	public String getOrgName() {
		return orgName;
	}

	public void setOrgName(String orgName) {
		this.orgName = orgName;
	}

	public String getUserId() {
		return userId;
	}

	public void setUserId(String userId) {
		this.userId = userId;
	}

	public List<String> getRoles() {
		return roles;
	}

	public void setRoles(List<String> roles) {
		this.roles = roles;
	}

	public List<String> getPermissions() {
		return permissions;
	}

	public void setPermissions(List<String> permissions) {
		this.permissions = permissions;
	}

}
